change file name without s
