+++ scopes/dependencies/modules/packages-excluder/index.ts
@@ -1,1 +1,1 @@
-export { generateNodeModulesPatterns } from './generate-node-modules-patterns';
+export { generateNodeModulesPattern } from './generate-node-modules-pattern';
