+++ scopes/dependencies/modules/packages-excluder/packages-excluder.docs.md
@@ -3,7 +3,7 @@ labels: ['typescript', 'utils', 'packages', 'node modules', 'node', 'regex', 'ex
 description: 'Create node modules regex with packages.'
 ---
 
-import { generateNodeModulesPatterns } from './generate-node-modules-patterns';
+import { generateNodeModulesPattern } from './generate-node-modules-pattern';
 
 A function that receive an array of packages name to catch in node modules and return a regex of it.  
 The returned regex can be used in Jest `transformIgnorePatterns` to ignore specific packages.
@@ -13,6 +13,6 @@ For example:
 ```js live
 () => {
   const packagesToTransform = ['react', '@myorg', 'testing-library__dom'];
-  return generateNodeModulesPatterns({ packages: packagesToTransform });
+  return generateNodeModulesPattern({ packages: packagesToTransform });
 };
 ```
